+++ src/whitenoise/database/rust_migrations/m0016_processed_events_pubkey_fk.rs
@@ -9,7 +9,7 @@ pub struct Migration;
 #[async_trait]
 impl GlobalMigration for Migration {
     fn version(&self) -> u32 {
-        14
+        16
     }
 
     fn description(&self) -> &'static str {
