+++ src/whitenoise/database/rust_migrations/m0017_move_account_settings.rs
@@ -9,7 +9,7 @@ pub struct Migration;
 #[async_trait]
 impl LocalMigration for Migration {
     fn version(&self) -> u32 {
-        15
+        17
     }
 
     fn description(&self) -> &'static str {
@@ -38,7 +38,7 @@ impl LocalMigration for Migration {
         // Copy this account's row from shared DB, if any. Tolerant of missing
         // shared.account_settings (fresh install, or table already dropped by
         // the post-local cleanup migration). If the source table is missing
-        // because v21 dropped it before this v15 ran for the current account
+        // because v23 dropped it before this v17 ran for the current account
         // (a sequencing bug — `MIGRATOR.run_all` at boot is supposed to make
         // this impossible), log a warn so the data loss isn't silent.
         let shared_table_exists: bool = sqlx::query_scalar(
@@ -50,10 +50,10 @@ impl LocalMigration for Migration {
 
         if !shared_table_exists {
             tracing::warn!(
-                target: "whitenoise::database::rust_migrations::m0015",
+                target: "whitenoise::database::rust_migrations::m0017",
                 account = account_pubkey,
                 "shared.account_settings is missing; skipping local copy. \
-                 Expected only on fresh installs or when v21 already dropped \
+                 Expected only on fresh installs or when v23 already dropped \
                  the table — flag this if it appears for an upgrading account."
             );
         }
