+++ src/whitenoise/database/rust_migrations/m0019_move_published_key_packages.rs
@@ -9,7 +9,7 @@ pub struct Migration;
 #[async_trait]
 impl LocalMigration for Migration {
     fn version(&self) -> u32 {
-        17
+        19
     }
 
     fn description(&self) -> &'static str {
